Updating the collison system to account for cats translations

+++ Ported/LabRat/Assets/DOTSPorted/DOTSScripts/ComponentData/GameState.cs
@@ -7,5 +7,5 @@ using UnityEngine;
 public struct GameState : IComponentData
 {
     public int timer;
-    public int collisionRadius;
+    public float collisionRadius;
 }
